refactor: function to turn transactions into payments
